Refactoring and bug fixes Using just waveshare library directly Fixed network interface settings Updated ReadMe
