--- waveshare_library/epdif.py
@@ -1,68 +1,0 @@
-##
- #  @filename   :   epdif.py
- #  @brief      :   EPD hardware interface implements (GPIO, SPI)
- #  @author     :   Yehui from Waveshare
- #
- #  Copyright (C) Waveshare     July 4 2017
- #
- # Permission is hereby granted, free of charge, to any person obtaining a copy
- # of this software and associated documnetation files (the "Software"), to deal
- # in the Software without restriction, including without limitation the rights
- # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- # copies of the Software, and to permit persons to  whom the Software is
- # furished to do so, subject to the following conditions:
- #
- # The above copyright notice and this permission notice shall be included in
- # all copies or substantial portions of the Software.
- #
- # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- # FITNESS OR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- # AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- # LIABILITY WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- # THE SOFTWARE.
- #
-
-import spidev
-import RPi.GPIO as GPIO
-import time
-
-# Pin definition
-RST_PIN         = 17
-DC_PIN          = 25
-CS_PIN          = 8
-BUSY_PIN        = 24
-
-# Display resolution
-EPD_WIDTH       = 176
-EPD_HEIGHT      = 264
-
-# SPI device, bus = 0, device = 0
-SPI = spidev.SpiDev(0, 0)
-
-def epd_digital_write(pin, value):
-    GPIO.output(pin, value)
-
-def epd_digital_read(pin):
-    return GPIO.input(BUSY_PIN)
-
-def epd_delay_ms(delaytime):
-    time.sleep(delaytime / 1000.0)
-
-def spi_transfer(data):
-    SPI.writebytes(data)
-
-def epd_init():
-    GPIO.setmode(GPIO.BCM)
-    GPIO.setwarnings(False)
-    GPIO.setup(RST_PIN, GPIO.OUT)
-    GPIO.setup(DC_PIN, GPIO.OUT)
-    GPIO.setup(CS_PIN, GPIO.OUT)
-    GPIO.setup(BUSY_PIN, GPIO.IN)
-
-    SPI.max_speed_hz = 2000000
-    SPI.mode = 0b00
-    return 0
-
-### END OF FILE ###
